urfu/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.2

